Android: reconcile UI text with editor text
This should help in cases where the UI text is modified externally by the OS or some app.

+++ platforms/android/library/src/main/java/io/element/android/wysiwyg/internal/viewmodel/EditorViewModel.kt
@@ -74,8 +74,8 @@ internal class EditorViewModel(
                     // This conversion to a plain String might be too simple
                     composer?.replaceTextIn(
                         action.value.toString(),
-                        action.start.toUInt(),
-                        action.end.toUInt()
+                        action.start,
+                        action.end
                     )
                 }
